SpeedChangeBehavior now independent of framerate

+++ src/de/tum/in/flowgame/strategy/FunctionStrategy2MBean.java
@@ -2,7 +2,7 @@ package de.tum.in.flowgame.strategy;
 
 public interface FunctionStrategy2MBean {
 
-	int getCurrentPosition();
+	long getCurrentPosition();
 	float getShortTerm();
 	float getMidTerm();
 	float getLongTerm();
